Renames player mod to voice.
