+++ src/app.rs
@@ -1,7 +1,7 @@
 use crate::database::migrate;
 use crate::intro::{get_url_for_user_and_guild, user_joined_voice};
-use crate::player::get_yt_track_duration;
-use crate::player::play;
+use crate::voice::get_yt_track_duration;
+use crate::voice::play;
 
 use poise::serenity_prelude as serenity;
 use serenity::{client::Client, prelude::GatewayIntents};
