+++ src/lib.rs
@@ -1,4 +1,4 @@
 pub mod app;
 pub mod database;
 pub mod intro;
-pub mod player;
+pub mod voice;
